requirements.txt
torch==1.4.0
torch-cluster==1.5.4
torch-geometric==1.5.0
torch-scatter==2.0.4
torch-sparse==0.6.0
scikit-learn==0.21.3
numpy==1.16.4
networkx==2.3
magic-impute==2.0.3
